==> utilities/2. Processing_and_store_MTCNN.ipynb <==
try:
    from google.colab import drive
    IN_COLAB = True
    print("Running on Google Colab. ")
except:
    IN_COLAB = False
    print("Not running on Google Colab. ")

if IN_COLAB:
    !pip install facenet-pytorch

from facenet_pytorch import MTCNN
import cv2
from PIL import Image
import numpy as np
from matplotlib import pyplot as plt
from tqdm.notebook import tqdm
import torch

# Create face detector
device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')
print('Running on device: {}'.format(device))
mtcnn = MTCNN(select_largest=False, device=device, image_size=224)

if IN_COLAB:
    from google.colab import drive
    drive.mount('/content/drive/')

if IN_COLAB:
    test_set_path = "/content/drive/Shareddrives/AI4CYBSEC/face_dataset/test_set"
else:
    test_set_path = "face_dataset/test_set"

# List of files in the dataset
import os
root_dir = test_set_path
file_list = []
for root, dirs, files in os.walk(root_dir):
    for file in files:
        file_list.append(os.path.join(root, file))
print(file_list[0])

list_folder = file_list[0].split("/")
print(list_folder[-1])
import os
# print the name of diurectory where the file is located
print(os.path.basename(os.path.dirname(file_list[0])))

import os

# Define the root folder and dataset name
if IN_COLAB:
    root_MTCNN_dataset = "/content/drive/Shareddrives/AI4CYBSEC/face_dataset"
else:
    root_MTCNN_dataset = "face_dataset"
name_dataset = "test_set_MTCNN_NN2"

# Combine the root folder and dataset name to create the full path
dataset_path = os.path.join(root_MTCNN_dataset, name_dataset)

# Create the folder if it doesn't exist
if not os.path.exists(dataset_path):
    os.makedirs(dataset_path)
    print("Folder created:", dataset_path)
else:
    print("Folder already exists.")


for i in range(len(file_list)):

  frame =Image.open(file_list[i])
  name_image = os.path.basename(file_list[i])
  labels = os.path.basename(os.path.dirname(file_list[i]))
  label_path = os.path.join(dataset_path,labels)
  if not os.path.exists(label_path):
    os.makedirs(label_path)
  path_save = os.path.join(label_path,name_image)
  print(f"{i}/{len(file_list)} -- Image name : {name_image} Labels: {labels} [PATH SAVE: {path_save}]" )

  mtcnn(frame, save_path=path_save)
